requirements.txt
altgraph==0.17.4
asgiref==3.11.1
attrs==26.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
distlib==0.4.0
Django==6.0.3
djangorestframework==3.17.1
docutils==0.22.2
drf-spectacular==0.29.0
filelock==3.20.2
filetype==1.2.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
packaging==25.0
pefile==2023.2.7
platformdirs==4.5.1
psycopg2-binary==2.9.11
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-decouple==3.8
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
setuptools==80.9.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.4
